⤴️ Patch gif link

+++ utils/automod.py
@@ -13,6 +13,7 @@ class AutoModUtils(commands.Cog):
             self.bad_words = json.load(bad_words_file)["bad_words"]
             self.bad_word_patterns = [re.compile(rf'\b{re.escape(word)}\b', re.IGNORECASE) for word in self.bad_words]
             self.link_pattern = re.compile(r'http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\\(\\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
+            self.tenor_link = re.compile(r'tenor.com')
 
     @commands.Cog.listener()
     async def on_ready(self):
@@ -43,23 +44,24 @@ class AutoModUtils(commands.Cog):
                     await log_channel.send(embed=embed)
             
             if self.link_pattern.search(message.content):
-                log_channel = self.bot.get_channel(config["LOG_ID"])
-                if log_channel:
-                    embed = disnake.Embed(title="Link Detected",
-                                        description=f"A link was fund from {message.author.mention}.",
-                                        color=disnake.Color.brand_red())
-                    embed.add_field(name="Contenu du Message", value=message.content, inline=False)
-                    await log_channel.send(embed=embed)
-                user_embed = disnake.Embed(title="Warning",
-                                        description=f"Your message ```{message.content}``` as been blocked.",
-                                        color=disnake.Color.brand_red())
-                await message.author.send(embed=user_embed)
-                await message.delete()
+                if not self.tenor_link.search(message.content):
+                    log_channel = self.bot.get_channel(config["LOG_ID"])
+                    if log_channel:
+                        embed = disnake.Embed(title="Link Detected",
+                                            description=f"A link was fund from {message.author.mention}.",
+                                            color=disnake.Color.brand_red())
+                        embed.add_field(name="Contenu du Message", value=message.content, inline=False)
+                        await log_channel.send(embed=embed)
+                    user_embed = disnake.Embed(title="Warning",
+                                            description=f"Your message ```{message.content}``` as been blocked.",
+                                            color=disnake.Color.brand_red())
+                    await message.author.send(embed=user_embed)
+                    await message.delete()
 
-                temp_embed = disnake.Embed(
-                    description=f"A link as been blocker.\nPlease don't send link {message.author.mention}",
-                    color=disnake.Color.brand_red())
-                await message.channel.send(embed=temp_embed, delete_after=10.0)
+                    temp_embed = disnake.Embed(
+                        description=f"A link as been blocker.\nPlease don't send link {message.author.mention}",
+                        color=disnake.Color.brand_red())
+                    await message.channel.send(embed=temp_embed, delete_after=10.0)
 
 def setup(bot):
     bot.add_cog(AutoModUtils(bot))
